+++ src/main/java/org/etmetmy/bn_server/domain/memo/entity/Memo.java
@@ -6,9 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.etmetmy.bn_server.domain.project.entity.Project;
+import org.etmetmy.bn_server.domain.user.entity.User;
 import org.etmetmy.bn_server.global.entity.BaseEntity;
-
-import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "memo")
@@ -29,4 +28,6 @@ public class Memo extends BaseEntity {
 
     @Column(name = "content", length = 1000)
     private String content;
+
+
 }
