+++ src/main/java/org/etmetmy/bn_server/domain/memo/service/MemoServiceImpl.java
@@ -1,9 +1,9 @@
-package org.etmetmy.bn_server.domain.memo.entity.service;
+package org.etmetmy.bn_server.domain.memo.service;
 
 import lombok.RequiredArgsConstructor;
 import org.etmetmy.bn_server.domain.memo.entity.Memo;
-import org.etmetmy.bn_server.domain.memo.entity.dto.entity.MemoResponse;
-import org.etmetmy.bn_server.domain.memo.entity.repository.MemoRepository;
+import org.etmetmy.bn_server.domain.memo.dto.entity.MemoResponse;
+import org.etmetmy.bn_server.domain.memo.repository.MemoRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,5 +18,11 @@ public class MemoServiceImpl implements MemoService{
     public List<MemoResponse> getProjectMemos(Long projectId) {
         List<Memo> memos = memoRepository.findByProjectId(projectId);
         return MemoResponse.Converter.from(memos);
+    }
+
+    @Override
+    public Memo getUserMemo() {
+        Memo memo = memoRepository.findByUserId()
+        return null;
     }
 }
