+++ src/main/java/org/etmetmy/bn_server/domain/user/controller/UserController.java
@@ -104,4 +104,6 @@ public class UserController {
         return CommonResponse.success("회원 정보 수정 완료",userId);
     }
 
+
+
 }
